+++ crates/c-api/src/lib.rs
@@ -1,0 +1,6 @@
+#![expect(clippy::missing_safety_doc)]
+pub mod basic_types;
+pub mod boxes;
+pub mod demux;
+pub mod error;
+pub mod mux;